add minimal working example of overlay
also requires:
- (being a privapp)
- granting runtime permission SYSTEM_ALERT_WINDOW via ADB, both for user 0 and 10 (current user)

+++ frontend/app/src/main/java/de/amosproj3/ziofa/ui/shared/ConfigurationConverters.kt
@@ -102,7 +102,7 @@ fun BackendFeatureOptions.setInLocalConfiguration(
                             fnName = this.method,
                             target = this.odexFilePath,
                             offset = this.offset,
-                            pid = it.toInt(),
+                            pid = it,
                         )
                     },
             )
